traverser: return during partial cancel when allocation exists
Problem: issue #1284 identified a scenario where rabbits are not
released due to a traverser error during partial cancellation. The
traverser should skip the rest of the mod_plan function when an
allocation is found and mod_data.mod_type ==
job_modify_t::PARTIAL_CANCEL.

Add a goto statement to return 0 under this circumstance.